chore: divers fixes on frontend
