+++ openaev-front/src/admin/components/scenarios/scenario/Scenario.tsx
@@ -42,13 +42,13 @@ import {
   type Scenario as ScenarioType,
   type SearchPaginationInput,
   type SortField,
+  type WorkflowScopeRuleOutput,
 } from '../../../../utils/api-types';
 import { useAppDispatch } from '../../../../utils/hooks';
 import useDataLoader from '../../../../utils/hooks/useDataLoader';
 import { AbilityContext } from '../../../../utils/permissions/permissionsContext';
 import { ACTIONS, SUBJECTS } from '../../../../utils/permissions/types';
 import { isEmptyField, isFeatureEnabled } from '../../../../utils/utils';
-import { isScopeMissingForChaining } from '../../common/healthchecks/scopeHealthcheckUtils';
 import MitreCoverageMatrix from '../../common/matrix/MitreCoverageMatrix';
 import ExercisePopover from '../../simulations/simulation/ExercisePopover';
 import SimulationList from '../../simulations/SimulationList';
@@ -91,13 +91,18 @@ const Scenario = ({ setOpenInstantiateSimulationAndStart }: { setOpenInstantiate
       ? helper.getWorkflowConfiguration(scenarioWorkflowId)
       : undefined,
   }));
+  const hasAllowlistEntry = (workflowScopeRules: WorkflowScopeRuleOutput[] = []): boolean =>
+    workflowScopeRules.some((rule: WorkflowScopeRuleOutput) =>
+      rule.workflow_scope_rule_selected_mode === 'ALLOWLIST'
+      && !!rule.workflow_scope_rule_value?.trim(),
+    );
 
   // Overview launch CTA follows the same allowlist-required rule as header/scope pages.
-  const isScopeMissing = isScopeMissingForChaining({
-    isChaining: isScenarioChaining,
-    workflowScopeRules: workflowConfiguration?.workflow_scope_rules ?? [],
-    healthchecks,
-  });
+  const isScopeMissing = isScenarioChaining
+    && (
+      !hasAllowlistEntry(workflowConfiguration?.workflow_scope_rules ?? [])
+      || healthchecks.some((hc: HealthCheck) => hc.type === ('SCOPE_DEFINITION' as HealthCheck['type']) && hc.detail === 'EMPTY')
+    );
 
   const agentsActive = useMemo(() => {
     const injectAssetIds: string[] = injects.flatMap((inject: Inject) => inject.inject_assets);
