+++ openaev-front/src/admin/components/scenarios/scenario/scope/ScenarioScope.tsx
@@ -5,12 +5,25 @@ import type { WorkflowConfigurationHelper } from '../../../../../actions/chainin
 import { searchScenarioHealthcheks } from '../../../../../actions/scenarios/scenario-actions';
 import type { ScenariosHelper } from '../../../../../actions/scenarios/scenario-helper';
 import { useHelper } from '../../../../../store';
-import { type HealthCheck, type Scenario } from '../../../../../utils/api-types';
+import { type HealthCheck, type Scenario, type WorkflowScopeRuleOutput } from '../../../../../utils/api-types';
 import ScopeDefinition from '../../../chaining/ScopeDefinition';
 import Healthchecks from '../../../common/healthchecks/Healthchecks';
-import { getScopeAwareHealthchecks, isScopeMissingForChaining } from '../../../common/healthchecks/scopeHealthcheckUtils';
 
 const ScenarioScope = () => {
+  const isScopeDefinitionEmptyHealthcheck = (healthcheck: HealthCheck): boolean =>
+    healthcheck.type === 'SCOPE_DEFINITION' && healthcheck.detail === 'EMPTY';
+  const hasAllowlistEntry = (workflowScopeRules: WorkflowScopeRuleOutput[] = []): boolean =>
+    workflowScopeRules.some((rule: WorkflowScopeRuleOutput) =>
+      rule.workflow_scope_rule_selected_mode === 'ALLOWLIST'
+      && !!rule.workflow_scope_rule_value?.trim(),
+    );
+  const SCOPE_DEFINITION_EMPTY_WARNING: HealthCheck = {
+    creation_date: '',
+    detail: 'EMPTY',
+    status: 'WARNING',
+    type: 'SCOPE_DEFINITION',
+  };
+
   const { scenarioId } = useParams() as { scenarioId: Scenario['scenario_id'] };
 
   const { scenario } = useHelper((helper: ScenariosHelper) => ({ scenario: helper.getScenario(scenarioId) }));
@@ -25,16 +38,22 @@ const ScenarioScope = () => {
   const [healthchecks, setHealthchecks] = useState<HealthCheck[]>([]);
   const isScenarioChaining = !!scenario?.scenario_workflow_id;
   // Scope page is chaining-only; keep state explicit so helpers stay readable.
-  const isScopeMissing = isScopeMissingForChaining({
-    isChaining: isScenarioChaining,
-    workflowScopeRules: workflowConfiguration?.workflow_scope_rules ?? [],
-    healthchecks,
-  });
-  const healthchecksForBanner = getScopeAwareHealthchecks({
-    healthchecks,
-    isChaining: isScenarioChaining,
-    isScopeMissing,
-  });
+  const isScopeMissing = isScenarioChaining
+    && (
+      !hasAllowlistEntry(workflowConfiguration?.workflow_scope_rules ?? [])
+      || healthchecks.some(isScopeDefinitionEmptyHealthcheck)
+    );
+  const healthchecksForBanner = (() => {
+    if (!isScenarioChaining) {
+      return healthchecks;
+    }
+    const withoutScopeDefinition = healthchecks.filter((healthcheck: HealthCheck) => healthcheck.type !== 'SCOPE_DEFINITION');
+    if (!isScopeMissing) {
+      return withoutScopeDefinition;
+    }
+    const scopeDefinitionHealthcheck = healthchecks.find(isScopeDefinitionEmptyHealthcheck) ?? SCOPE_DEFINITION_EMPTY_WARNING;
+    return [...withoutScopeDefinition, scopeDefinitionHealthcheck];
+  })();
 
   useEffect(() => {
     searchScenarioHealthcheks(scenarioId).then((result: { data: HealthCheck[] }) => setHealthchecks(result.data));
