+++ openaev-front/src/admin/components/simulations/simulation/scope/SimulationScope.tsx
@@ -5,12 +5,25 @@ import type { WorkflowConfigurationHelper } from '../../../../../actions/chainin
 import { searchExerciseHealthchecks } from '../../../../../actions/Exercise';
 import type { ExercisesHelper } from '../../../../../actions/exercises/exercise-helper';
 import { useHelper } from '../../../../../store';
-import { type Exercise, type HealthCheck } from '../../../../../utils/api-types';
+import { type Exercise, type HealthCheck, type WorkflowScopeRuleOutput } from '../../../../../utils/api-types';
 import ScopeDefinition from '../../../chaining/ScopeDefinition';
 import Healthchecks from '../../../common/healthchecks/Healthchecks';
-import { getScopeAwareHealthchecks, isScopeMissingForChaining } from '../../../common/healthchecks/scopeHealthcheckUtils';
 
 const SimulationScope = () => {
+  const isScopeDefinitionEmptyHealthcheck = (healthcheck: HealthCheck): boolean =>
+    healthcheck.type === 'SCOPE_DEFINITION' && healthcheck.detail === 'EMPTY';
+  const hasAllowlistEntry = (workflowScopeRules: WorkflowScopeRuleOutput[] = []): boolean =>
+    workflowScopeRules.some((rule: WorkflowScopeRuleOutput) =>
+      rule.workflow_scope_rule_selected_mode === 'ALLOWLIST'
+      && !!rule.workflow_scope_rule_value?.trim(),
+    );
+  const SCOPE_DEFINITION_EMPTY_WARNING: HealthCheck = {
+    creation_date: '',
+    detail: 'EMPTY',
+    status: 'WARNING',
+    type: 'SCOPE_DEFINITION',
+  };
+
   const { exerciseId } = useParams() as { exerciseId: Exercise['exercise_id'] };
 
   const { exercise } = useHelper((helper: ExercisesHelper) => ({ exercise: helper.getExercise(exerciseId) }));
@@ -25,16 +38,22 @@ const SimulationScope = () => {
   const [healthchecks, setHealthchecks] = useState<HealthCheck[]>([]);
   const isSimulationChaining = !!exercise?.exercise_workflow_id;
   // Scope page is chaining-only; keep state explicit so helpers stay readable.
-  const isScopeMissing = isScopeMissingForChaining({
-    isChaining: isSimulationChaining,
-    workflowScopeRules: workflowConfiguration?.workflow_scope_rules ?? [],
-    healthchecks,
-  });
-  const healthchecksForBanner = getScopeAwareHealthchecks({
-    healthchecks,
-    isChaining: isSimulationChaining,
-    isScopeMissing,
-  });
+  const isScopeMissing = isSimulationChaining
+    && (
+      !hasAllowlistEntry(workflowConfiguration?.workflow_scope_rules ?? [])
+      || healthchecks.some(isScopeDefinitionEmptyHealthcheck)
+    );
+  const healthchecksForBanner = (() => {
+    if (!isSimulationChaining) {
+      return healthchecks;
+    }
+    const withoutScopeDefinition = healthchecks.filter((healthcheck: HealthCheck) => healthcheck.type !== 'SCOPE_DEFINITION');
+    if (!isScopeMissing) {
+      return withoutScopeDefinition;
+    }
+    const scopeDefinitionHealthcheck = healthchecks.find(isScopeDefinitionEmptyHealthcheck) ?? SCOPE_DEFINITION_EMPTY_WARNING;
+    return [...withoutScopeDefinition, scopeDefinitionHealthcheck];
+  })();
 
   useEffect(() => {
     if (exercise?.exercise_workflow_id) {
